refactor: simplify i18n config hooks
