+++ packages/react-core/src/context.ts
@@ -15,7 +15,7 @@ export {
   useCustomMapping,
   useDefaultLocale,
   useLocales,
-} from './hooks/external-store-hooks';
+} from './hooks/i18n-config';
 export { useGT } from './hooks/useGT';
 export { useMessages } from './hooks/useMessages';
 export { useTranslations } from './hooks/useTranslations';
